add フェーズによってファイル分け
