+++ src/app/Finish.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Finish = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Finish
