+++ src/app/Prologue.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Prologue = () => {
+  return (
+    <div>
+      
+    </div>
+  )
+}
+
+export default Prologue
